proc.c: adjust argument type
* proc.c (proc_new): adjust `is_lambda` argument type to
  `rb_vm_make_proc_lambda` to suppress a warning by VC6.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@48420 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ proc.c
@@ -559,7 +559,7 @@ bind_receiver(VALUE bindval)
 }
 
 static VALUE
-proc_new(VALUE klass, int is_lambda)
+proc_new(VALUE klass, int8_t is_lambda)
 {
     VALUE procval = Qnil;
     rb_thread_t *th = GET_THREAD();
